+++ README.md
@@ -113,7 +113,7 @@ cd src && mkdir shared && mkdir config
   - 지도에 보여질 emoji 스티커
   - `spot`에 대한 emoticon 스티커 개념
 - `place/`
-  - `네이버 API`를 통해 받아올 장소에 대한 정보
+  - `카카오 지역검색 API`를 통해 받아올 장소에 대한 정보
   - 매번 유저가 쿼리를 보낼 때마다, api로 장소 데이터를 받아오는게 비효율적이라면, 한번 요청된 데이터는 캐싱한다. (mongo 재활용 또는 redis 사용)
 - `shared/`
   - 각 domain들이 공통으로 사용할 util
