+++ src/place/place.module.ts
@@ -1,7 +1,7 @@
 import { Module } from "@nestjs/common";
 import { ConfigModule } from "../config/config.module";
 
-import { SearchService } from "./kakaoMap/search.service";
+import { SearchService } from "./kakaoMapSearch/search.service";
 import { PlaceResolver } from "./place.resolver";
 
 @Module({
